requirements.txt
# Django and REST framework
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# File handling
python-magic>=0.4.27

# Development tools
python-dotenv>=1.0.0 